remove unneeded component
