+++ packages/shared/src/components/atoms/ProposalsDetailsButton.svelte
@@ -1,10 +1,36 @@
 <script lang="typescript">
-    import { Modal, ProposalsDetailsMenu, MeatballMenuButton } from '@ui'
+    import { Modal, MeatballMenuButton, MenuItem } from '@ui'
+    import { Icon } from '@auxiliary/icon'
+    import { openPopup, PopupId } from '../../../../desktop/lib/auxiliary/popup'
+    import { localize } from '@core/i18n'
 
-    let modal: Modal
+    export let modalPosition:
+        | {
+              top?: string
+              bottom?: string
+              left?: string
+              right?: string
+          }
+        | undefined = undefined
+
+    let modal: Modal | undefined
+
+    function onRevoteClick(): void {
+        openPopup({
+            id: PopupId.Revote,
+        })
+        modal?.close()
+    }
 </script>
 
-<div class="max-h-7 max-w-9 flex-none flex-initial overflow-visible relative">
-    <MeatballMenuButton onClick={modal?.toggle} />
-    <ProposalsDetailsMenu bind:modal position={{ right: '0' }} classes="mt-1.5" />
-</div>
+<MeatballMenuButton onClick={() => modal?.toggle()} />
+<Modal bind:this={modal} position={modalPosition} {...$$restProps}>
+    <div class="flex flex-col">
+        <MenuItem
+            icon={Icon.Delete}
+            iconProps={{ width: '16', height: '19' }}
+            title={localize('actions.revote')}
+            onClick={onRevoteClick}
+        />
+    </div>
+</Modal>
