+++ packages/shared/src/components/modals/index.js
@@ -3,5 +3,4 @@ export { default as CollectibleDetailsMenu } from './CollectibleDetailsMenu.svel
 export { default as ExpirationTimePickerMenu } from './ExpirationTimePickerMenu.svelte'
 export { default as Modal } from './Modal.svelte'
 export { default as NodeActionsMenu } from './NodeActionsMenu.svelte'
-export { default as ProposalsDetailsMenu } from './ProposalsDetailsMenu.svelte'
 export { default as RecipientAccountSelector } from './RecipientAccountSelector.svelte'
